+++ Shared/Extensions/JellyfinAPIExtensions/APISortOrderExtensions.swift
@@ -1,0 +1,26 @@
+//
+// Swiftfin is subject to the terms of the Mozilla Public
+// License, v2.0. If a copy of the MPL was not distributed with this
+// file, you can obtain one at https://mozilla.org/MPL/2.0/.
+//
+// Copyright (c) 2022 Jellyfin & Jellyfin Contributors
+//
+
+import Foundation
+import JellyfinAPI
+
+extension APISortOrder {
+    // TODO: Localize
+    var localized: String {
+        switch self {
+        case .ascending:
+            return "Ascending"
+        case .descending:
+            return "Descending"
+        }
+    }
+
+    var filter: ItemFilters.Filter {
+        .init(displayName: localized, filterName: rawValue)
+    }
+}
